StringReverse
Used a stack to pass in a string and reverse each word in the string. For instance, if you have the string “When Chuck Norris falls in water, Chuck Norris doesn't get wet. Water gets Chuck Norris.” It should now say, “nehW kcuhC sirroN sllaf ni retaw, …” That is, it should take in one string and be a function (or multiple functions) that use a stack (or multiple stacks) to reverse each word in the sentence. Created a text file got Chuck Norris jokes and have the program reverse each of the jokes.